fix: solve company id type inconsistency
#15

+++ backend/src/main/java/com/camyo/backend/empresa/EmpresaRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
 
     @Query("SELECT DISTINCT empresa FROM Empresa empresa WHERE empresa.usuario.id = :usuarioId")
-	public Optional<Empresa> obtenerPorUsuario(Long usuarioId);
+	public Optional<Empresa> obtenerPorUsuario(int usuarioId);
 
 }
